RLAI RUFILAR, The LFO Rising
The LFO finds itself out of touch and irrelevant to today's generation. This is an effort to bring the LFO back to serving its flock under God's Umbrella. The Third and Fourth Generation Lardizabals pull together after realizing there is so much good needed to do in this world. Having strong faith-based values, the clan sets out to once more become integral to the community built by Rufino Lardizabal over a century ago.